Simplified installtion
Now we use the 'deamon' user and 'nogroup' group to make installation easier on folks. I don't really want to set up a new user to install this thing, and I'm sure most people are the same.

This will now be rolled into the licensing tarball we have in the downloads section of the site.