Initial Commit of Draft Watchfolder Example
This is a python script designed to run constantly in the background watching
a folder for new additions. It is designed to create Draft jobs and submit
them to Deadline automatically. It doesn't do anything with the initial file
after the job is created, but it could be easily moved or deleted.